move scripts test ports to bin file
