+++ packages/scripts/src/helpers/config.ts
@@ -1,60 +1,12 @@
 import ipPkg from 'ip';
 import {
     toBoolean, toSafeString, isCI,
-    toIntegerOrThrow, isPortInUse, getAvailablePort
+    toIntegerOrThrow
 } from '@terascope/utils';
 import { Service } from './interfaces.js';
 import { kafkaVersionMapper } from './mapper.js';
-import signale from './signale.js';
 
 const { address } = ipPkg;
-
-// Validate that all ports aren't in use
-const allPorts = {
-    TEST_ELASTICSEARCH: {
-        ELASTICSEARCH_PORT: process.env.ELASTICSEARCH_PORT || '49200'
-    },
-    TEST_RESTRAINED_ELASTICSEARCH: {
-        RESTRAINED_ELASTICSEARCH_PORT: process.env.RESTRAINED_ELASTICSEARCH_PORT || '49202'
-    },
-    TEST_KAFKA: {
-        KAFKA_PORT: process.env.KAFKA_PORT || '49092',
-        ZOOKEEPER_CLIENT_PORT: process.env.ZOOKEEPER_CLIENT_PORT || '42181',
-    },
-    TEST_MINIO: {
-        MINIO_PORT: process.env.MINIO_PORT || '49000',
-        MINIO_UI_PORT: process.env.MINIO_UI_PORT || '49001'
-    },
-    TEST_RABBITMQ: {
-        RABBITMQ_PORT: process.env.RABBITMQ_PORT || '45672',
-        RABBITMQ_MANAGEMENT_PORT: process.env.RABBITMQ_MANAGEMENT_PORT || '55672',
-    },
-    TEST_OPENSEARCH: {
-        OPENSEARCH_PORT: process.env.OPENSEARCH_PORT || '49210'
-    },
-    TEST_RESTRAINED_OPENSEARCH: {
-        RESTRAINED_OPENSEARCH_PORT: process.env.RESTRAINED_OPENSEARCH_PORT || '49206'
-    }
-};
-
-for (const [testKey, portsObj] of Object.entries(allPorts)) {
-    const serviceEnabled = toBoolean(process.env[testKey]);
-
-    if (serviceEnabled) {
-        for (const [envVarName, port] of Object.entries(portsObj)) {
-            signale.debug(`Checking availability of port ${port}`);
-
-            if (await isPortInUse(Number.parseInt(port))) {
-                signale.warn(`port ${port} is in use. Switching port..`);
-                const newPort = await getAvailablePort();
-                process.env[envVarName] = newPort.toString();
-                signale.warn(`${envVarName} env variable now uses port ${newPort}`);
-            } else {
-                signale.debug(`port ${port} is valid for env variable ${envVarName}`);
-            }
-        }
-    }
-}
 
 const forceColor = process.env.FORCE_COLOR || '1';
 export const FORCE_COLOR = toBoolean(forceColor)
@@ -95,7 +47,7 @@ export const RESTRAINED_ELASTICSEARCH_HOST = `http://${ELASTICSEARCH_HOSTNAME}:$
 
 export const KAFKA_NAME = process.env.KAFKA_NAME || 'kafka';
 export const KAFKA_HOSTNAME = process.env.KAFKA_HOSTNAME || HOST_IP;
-export const KAFKA_PORT = process.env.KAFKA_PORT || '49092';
+export const KAFKA_PORT = process.env.KAFKA_PORT || '49094';
 export const KAFKA_BROKER = `${KAFKA_HOSTNAME}:${KAFKA_PORT}`;
 export const KAFKA_VERSION = process.env.KAFKA_VERSION || '3.7.2';
 export const KAFKA_DOCKER_IMAGE = process.env.KAFKA_DOCKER_IMAGE || 'confluentinc/cp-kafka';
